+++ CustomScripts.cs
@@ -35,7 +35,7 @@ internal static partial class CustomScripts
         { "Relic_Consumable_ProfaneHeart", ProfaneHeart },
         { "Ring_DowngradedRing", DowngradedRing },
         { "Ring_BandOfTheFanatic", BandOfTheFanatic },
-        //Weapon_Anguish // TODO: Has to be injected if we want to support it
+        { "Weapon_Anguish", Anguish },
         //Amulet_ParticipationMedal // TODO: Has to be injected if we want to support it
         { "Weapon_Deceit", lic => { Deceit(lic); return true; } },
         { "Weapon_Godsplitter", lic => { Godsplitter(lic); return true; } },